requirements.txt
anyio==3.6.2
ariadne==0.17.1
click==8.1.3
gitdb==4.0.10
GitPython==3.1.30
graphql-core==3.2.3
h11==0.14.0
idna==3.4
smmap==5.0.0
sniffio==1.3.0
starlette==0.23.1
typing_extensions==4.4.0
uvicorn==0.20.0
